optimize footnote plugin
