+++ src/lib/footnotes-utils.ts
@@ -18,13 +18,8 @@ import { FootnoteWithIndex } from '@manuscripts/style-guide'
 import { skipTracking } from '@manuscripts/track-changes-plugin'
 import { isFootnoteNode, ManuscriptNode, schema } from '@manuscripts/transform'
 import { Fragment } from 'prosemirror-model'
-import { EditorState, Transaction } from 'prosemirror-state'
-import { ReplaceStep } from 'prosemirror-transform'
-import {
-  findChildrenByType,
-  findParentNodeClosestToPos,
-  NodeWithPos,
-} from 'prosemirror-utils'
+import { Transaction } from 'prosemirror-state'
+import { findChildrenByType, NodeWithPos } from 'prosemirror-utils'
 
 export const getNewFootnotePos = (
   footnotesElement: NodeWithPos,
@@ -105,71 +100,5 @@ export const orderTableFootnotes = (
       tr.mapping.map(footnoteElementPos + footnotesElement.nodeSize),
       Fragment.fromArray(orderedFootnotes)
     )
-  )
-}
-
-export const updateTableInlineFootnoteLabels = (
-  transactions: readonly Transaction[],
-  oldState: EditorState,
-  newState: EditorState
-) => {
-  const tableInlineFootnoteChange = transactions.find((tr) =>
-    tr.steps.find((s) => {
-      if (s instanceof ReplaceStep) {
-        const $pos = oldState.doc.resolve((s as ReplaceStep).from)
-        return (
-          $pos.node().type === schema.nodes.table_cell &&
-          $pos.node($pos.depth - 2).type === schema.nodes.table
-        )
-      }
-    })
-  )
-
-  if (!tableInlineFootnoteChange) {
-    return null
-  }
-
-  const step = tableInlineFootnoteChange.steps[0] as ReplaceStep
-
-  const tr = newState.tr
-  const $pos = newState.doc.resolve(step.from)
-  const table = findParentNodeClosestToPos(
-    $pos,
-    (node) => node.type === schema.nodes.table_element
-  )
-
-  const footnotesElementWithPos =
-    table &&
-    findChildrenByType(table.node, schema.nodes.footnotes_element).pop()
-
-  if (!table || !footnotesElementWithPos) {
-    return null
-  }
-
-  const labels = buildTableFootnoteLabels(table.node)
-
-  findChildrenByType(table.node, schema.nodes.inline_footnote).map(
-    ({ node, pos }) => {
-      const contents = node.attrs.rids
-        .map((rid: string) => labels.get(rid))
-        .join(',')
-
-      if (
-        contents !== node.attrs.contents &&
-        tr.doc.nodeAt(tr.mapping.map(table.pos + pos + 1))
-      ) {
-        tr.setNodeMarkup(tr.mapping.map(table.pos + pos + 1), undefined, {
-          ...node.attrs,
-          rids: node.attrs.rids,
-          contents,
-        })
-      }
-    }
-  )
-
-  return orderTableFootnotes(
-    tr,
-    footnotesElementWithPos,
-    tr.mapping.map(table.pos)
   )
 }
